added view tests for monolith and accounts

+++ accounts_microservice/accounts/models.py
@@ -17,7 +17,7 @@ class State(models.Model):
 class Owner(models.Model):
     name = models.CharField(max_length=200)
     email = models.EmailField()
-    phone = models.PositiveSmallIntegerField(validators=[MaxValueValidator(9999999999)])
+    phone = models.PositiveBigIntegerField(validators=[MaxValueValidator(9999999999)])
     description = models.TextField(max_length=1000)
     account_number = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     state = models.ForeignKey(State, related_name="+", on_delete=models.PROTECT)
